pyproject.toml
[tool.pytest.ini_options]
pythonpath = "."

[tool.poetry]
name = "automatic-essay-evaluator"
version = "0.1.0"
description = "Automated tool for evaluation of natural language texts."
authors = ["Danis Alukaev <d.alukaev@innopolis.university>"]
license = "MIT"
readme = "README.md"
packages = [{include = "automatic_essay_evaluator"}]

[tool.poetry.dependencies]
python = "3.7.13"
flake8 = "^5.0.4"
hydra-core = "^1.3.0"
python-dotenv = "^0.21.0"
coverage = "^6.5.0"
streamlit = "^1.16.0"
transformers = {extras = ["sentencepiece"], version = "^4.25.1"}
sphinx = ">=4.0"
sphinx-rtd-theme = "^1.1.1"
sphinx-autoapi = "^2.0.0"
mypy = "^0.991"
types-requests = "^2.28.11.5"
pytest-cov = "^4.0.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
hydra-core = "^1.3.0"
isort = "^5.11.3"
kaggle = "^1.5.12"
numpy = "^1.21.6"
pre-commit = "^2.20.0"
pytest = "^7.2.0"
pytorch-lightning = "^1.8.5.post0"
tqdm = "^4.64.1"
torch = "^1.13.1"
transformers = "^4.25.1"
pyspellchecker = "^0.7.1"
catboost = "^1.1.1"
scikit-learn = "^1.0.2"
wandb = "^0.13.7"
easydict = "^1.10"
nltk = "^3.8"
seaborn = "^0.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
